extracted character classes into characters.py
